relcheck_v3/hallucination_generation/requirements.txt
openai>=1.0
pydantic>=2.0
python-Levenshtein
tqdm
pandas
Pillow
hypothesis
pytest
pytest-mock
